Server structure + support for taxeringsenhet typkoder tools
This structure allows us to define multiple paths with mcp servers using
the same tools. We could even merge in the same set of tools multiple
times into each server, as long as we define a prefix.

Ref: https://gofastmcp.com/servers/composition#mounting-live-linking

TODO: we'll probably need to move resources to someplace else, or rename
tools to be something else, since they contain multiple things.